requirements.txt
django==4.2.18
django-filter==23.1.0
dj-database-url==2.1.0
dj-static==0.0.6
python-decouple==3.8
django-environ==0.11.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0

djangorestframework==3.16.0 #追加
djangorestframework-simplejwt==5.5.0 #追加

